+++ com.taken_seat.payment-service/src/main/java/com/taken_seat/payment_service/application/dto/controller/request/PaymentRegisterReqDto.java
@@ -1,4 +1,4 @@
-package com.taken_seat.payment_service.application.dto.request;
+package com.taken_seat.payment_service.application.dto.controller.request;
 
 import java.util.UUID;
