+++ com.taken_seat.payment-service/src/main/java/com/taken_seat/payment_service/application/service/PaymentService.java
@@ -3,9 +3,9 @@ package com.taken_seat.payment_service.application.service;
 import java.util.UUID;
 
 import com.taken_seat.common_service.dto.AuthenticatedUser;
-import com.taken_seat.payment_service.application.dto.request.PaymentSearchReqDto;
-import com.taken_seat.payment_service.application.dto.response.PagePaymentResponseDto;
-import com.taken_seat.payment_service.application.dto.response.PaymentDetailResDto;
+import com.taken_seat.payment_service.application.dto.controller.request.PaymentSearchReqDto;
+import com.taken_seat.payment_service.application.dto.controller.response.PagePaymentResponseDto;
+import com.taken_seat.payment_service.application.dto.controller.response.PaymentDetailResDto;
 import com.taken_seat.payment_service.application.dto.service.PaymentDto;
 
 public interface PaymentService {
