+++ com.taken_seat.payment-service/src/main/java/com/taken_seat/payment_service/domain/repository/PaymentQuerydslRepository.java
@@ -2,7 +2,7 @@ package com.taken_seat.payment_service.domain.repository;
 
 import org.springframework.data.domain.Page;
 
-import com.taken_seat.payment_service.application.dto.request.PaymentSearchReqDto;
+import com.taken_seat.payment_service.application.dto.controller.request.PaymentSearchReqDto;
 import com.taken_seat.payment_service.domain.model.Payment;
 
 public interface PaymentQuerydslRepository {
