+++ com.taken_seat.payment-service/src/main/java/com/taken_seat/payment_service/infrastructure/repository/PaymentQuerydslRepositoryImpl.java
@@ -23,7 +23,7 @@ import com.querydsl.core.types.dsl.PathBuilder;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import com.taken_seat.common_service.exception.customException.PaymentException;
 import com.taken_seat.common_service.exception.enums.ResponseCode;
-import com.taken_seat.payment_service.application.dto.request.PaymentSearchReqDto;
+import com.taken_seat.payment_service.application.dto.controller.request.PaymentSearchReqDto;
 import com.taken_seat.payment_service.domain.enums.PaymentStatus;
 import com.taken_seat.payment_service.domain.model.Payment;
 import com.taken_seat.payment_service.domain.model.QPayment;
