+++ com.taken_seat.payment-service/src/main/java/com/taken_seat/payment_service/presentation/controller/PaymentController.java
@@ -18,11 +18,11 @@ import com.taken_seat.common_service.aop.annotation.RoleCheck;
 import com.taken_seat.common_service.aop.vo.Role;
 import com.taken_seat.common_service.dto.ApiResponseData;
 import com.taken_seat.common_service.dto.AuthenticatedUser;
-import com.taken_seat.payment_service.application.dto.request.PaymentRegisterReqDto;
-import com.taken_seat.payment_service.application.dto.request.PaymentSearchReqDto;
-import com.taken_seat.payment_service.application.dto.request.PaymentUpdateReqDto;
-import com.taken_seat.payment_service.application.dto.response.PagePaymentResponseDto;
-import com.taken_seat.payment_service.application.dto.response.PaymentDetailResDto;
+import com.taken_seat.payment_service.application.dto.controller.request.PaymentRegisterReqDto;
+import com.taken_seat.payment_service.application.dto.controller.request.PaymentSearchReqDto;
+import com.taken_seat.payment_service.application.dto.controller.request.PaymentUpdateReqDto;
+import com.taken_seat.payment_service.application.dto.controller.response.PagePaymentResponseDto;
+import com.taken_seat.payment_service.application.dto.controller.response.PaymentDetailResDto;
 import com.taken_seat.payment_service.application.dto.service.PaymentDto;
 import com.taken_seat.payment_service.application.service.PaymentService;
 import com.taken_seat.payment_service.infrastructure.mapper.PaymentMapper;
