+++ com.taken_seat.payment-service/src/test/java/com/taken_seat/payment_service/unit/PaymentServiceTest.java
@@ -22,9 +22,9 @@ import org.springframework.data.domain.PageRequest;
 
 import com.taken_seat.common_service.dto.AuthenticatedUser;
 import com.taken_seat.common_service.exception.customException.PaymentException;
-import com.taken_seat.payment_service.application.dto.request.PaymentSearchReqDto;
-import com.taken_seat.payment_service.application.dto.response.PagePaymentResponseDto;
-import com.taken_seat.payment_service.application.dto.response.PaymentDetailResDto;
+import com.taken_seat.payment_service.application.dto.controller.request.PaymentSearchReqDto;
+import com.taken_seat.payment_service.application.dto.controller.response.PagePaymentResponseDto;
+import com.taken_seat.payment_service.application.dto.controller.response.PaymentDetailResDto;
 import com.taken_seat.payment_service.application.dto.service.PaymentDto;
 import com.taken_seat.payment_service.application.service.PaymentServiceImpl;
 import com.taken_seat.payment_service.domain.enums.PaymentStatus;
